pyproject.toml
[tool.poetry]
name = "medion-technical-task"
version = "0.1.1"
description = "Техническое задание"
authors = ["Vladislav <rgvr2017@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.10"
django = "5.0.7"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
django-environ = "0.11.2"
psycopg2-binary = "2.9.9"
drf-spectacular = "0.27.2"
drf-spectacular-sidecar = "2024.7.1"
django-filter = "24.3"
black = "24.10.0"
gunicorn = "21.0.0"
pytest-django = "4.9.0"
dj-database-url = "2.3.0"
six = "1.16.0"
langid = "^1.1.6"
pillow = "10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
